Utility\CaseInsensitiveDictionaryTest: split `getAll()` specific tests to dedicated test class
