+++ tests/Utility/CaseInsensitiveDictionary/CaseInsensitiveDictionaryTest.php
@@ -212,29 +212,4 @@ class CaseInsensitiveDictionaryTest extends TestCase {
 			break;
 		}
 	}
-
-	/**
-	 * Test retrieving all data as recorded in the dictionary.
-	 *
-	 * Take note of the key changes!
-	 *
-	 * @covers ::getAll
-	 *
-	 * @return void
-	 */
-	public function testGetAll() {
-		$expected = [
-			'upper case'  => 'Uppercase key',
-			'proper case' => 'First char in caps in key',
-			'lower case'  => 'Lowercase key',
-			''            => 'Null key will be converted to empty string',
-			0             => 'false key will become integer 0 key',
-			1             => 'true key will become integer 1 key',
-			5             => 'Float key will be converted to integer key (cut off)',
-			100           => 'Explicit integer numeric key',
-		];
-
-		$dictionary = new CaseInsensitiveDictionary(self::DATASET);
-		$this->assertSame($expected, $dictionary->getAll());
-	}
 }
